CRUD on products

+++ src/app/Day2/productlist/productlist.component.ts
@@ -15,21 +15,18 @@ import { ProductService } from 'src/app/shared/product.service';
 export class ProductlistComponent implements OnInit {
   @Output() productClicked: EventEmitter<Product> = new EventEmitter();
   sCategory: any;
+  selected = 'user';
   title: string = 'Product List';
   msg: string;
   productData: Product[] = [];
   temp2 = this.productData;
   switch: boolean = false;
   button: string = 'Add Product';
-  display = new FormControl('', Validators.required);
-  file_store: FileList;
-  file_list: Array<string> = [];
+  c: number = 0;
   res: any;
-  addForm: any;
-  error: boolean
+  error: boolean;
 
   constructor(
-    private fb: FormBuilder,
     private pService: ProductService,
     private logger: LoggingService,
     private snackBar: MatSnackBar,
@@ -37,7 +34,6 @@ export class ProductlistComponent implements OnInit {
   ) {}
 
   ngOnInit(): void {
-    this.initLoginForm();
     this.pService.getProducts().subscribe((data) => {
       data.forEach((product) => {
         this.productData.push(product);
@@ -45,71 +41,53 @@ export class ProductlistComponent implements OnInit {
     });
   }
 
-  initLoginForm() {
-    this.addForm = this.fb.group({
-      name: ['', Validators.required],
-      price: ['', Validators.required],
-      category: ['', Validators.required],
-      rating: ['', Validators.required],
-    });
-    // this.addForm.value
-  }
-
-  isPristine(): boolean {
-    return this.addForm.pristine;
-  }
-  isTouched(): boolean {
-    return this.addForm.touched;
-  }
-  isValid(): boolean {
-    return this.addForm.valid;
-  }
-
   add() {
-    this.switch = !this.switch;
-    if (this.switch) {
-      this.button = 'Show Table';
-    } else {
-      this.button = 'Add Product';
-    }
-  }
+    const dialogRef = this.dialog.open(DialogComponent, {
+      data: { button: 'Add Product' },
+    });
 
-  getRandomInt(max) {
-    return Math.floor(Math.random() * max);
-  }
-
-  create() {
-    if (this.isTouched() && this.isValid()) {
-      let cat: category;
-      if (this.addForm.value.category == 'clothing') {
-        cat = category.clothing;
-      } else if (this.addForm.value.category == 'electronics') {
-        cat = category.electronics;
-      } else if (this.addForm.value.category == 'decor') {
-        cat = category.decor;
-      } else {
-        cat = category.grocery;
-      }
-      let product: Product = {
-        id: this.getRandomInt(1000),
-        name: this.addForm.value.name,
-        price: parseInt(this.addForm.value.price),
-        category: cat,
-        rating: parseInt(this.addForm.value.rating),
-        image: `assets/images/${this.file_store[0].name}`,
-      };
-      this.pService.createProduct(product).subscribe((data) => {
-        this.productData.push(data);
-        this.snackBar.open(
-          'New Product is created successfully with id ' + data.id,
-          'close',
-          {
-            duration: 2000,
-            // panelClass: ['blue-snackbar']
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        for(let product of this.productData) {
+          if (result.id == product.id) {
+            this.c -= 1
+          } else {
+            this.c += 1
           }
-        );
-      });
-    }
+        };
+        if(this.c == (this.productData.length - 2)) {
+          this.snackBar.open(
+            'Something went wrong your product is not added yet please try again',
+            'close',
+            {
+              duration: 2000,
+              // panelClass: ['blue-snackbar']
+            }
+          );
+          this.add();
+          this.c = 0
+        }
+        if (this.c == this.productData.length) {
+          this.pService.createProduct(result).subscribe((data) => {
+            this.productData.push(data);
+            this.snackBar.open(
+              'New Product is created successfully with id ' + data.id,
+              'close',
+              {
+                duration: 2000,
+                // panelClass: ['blue-snackbar']
+              }
+            );
+          });
+          this.c = 0
+        }
+      } else {
+        this.snackBar.open("Couldn't create new Product", 'close', {
+          duration: 2000,
+          // panelClass: ['blue-snackbar']
+        });
+      }
+    });
   }
 
   log() {
@@ -136,27 +114,46 @@ export class ProductlistComponent implements OnInit {
   }
 
   deleteProduct(id: number) {
-    this.pService.deleteProduct(id).subscribe(() => {
-      let fIndex = this.productData.findIndex((item) => item.id == id);
-      if (fIndex > -1) {
-        this.productData.splice(fIndex, 1);
-      }
-    });
+    if(confirm(`Are you sure you want to delete this product`)){
+      this.pService.deleteProduct(id).subscribe(() => {
+        let fIndex = this.productData.findIndex((item) => item.id == id);
+        if (fIndex > -1) {
+          this.productData.splice(fIndex, 1);
+          this.snackBar.open(
+            `Product with id ${id} is deleted successfully.`,
+            'close',
+            {
+              duration: 2000,
+              // panelClass: ['blue-snackbar']
+            }
+          );
+        }
+      });
+    }
   }
 
   updateProduct(products: Product) {
     const dialogRef = this.dialog.open(DialogComponent, {
-      data: { product: products },
+      data: { product: products, button: 'update' },
     });
 
     dialogRef.afterClosed().subscribe((result) => {
       console.log('The dialog was closed', result);
-      this.pService.updateProduct(result).subscribe(() => {
-        let fIndex = this.productData.findIndex((item) => item.id == result.id);
-        if (fIndex > -1) {
-          this.productData[fIndex] = result;
-        }
-      });
+      if (result) {
+        this.pService.updateProduct(result).subscribe(() => {
+          let fIndex = this.productData.findIndex(
+            (item) => item.id == result.id
+          );
+          if (fIndex > -1) {
+            this.productData[fIndex] = result;
+          }
+        });
+      } else {
+        this.snackBar.open('Operation was canceled', 'close', {
+          duration: 2000,
+          // panelClass: ['blue-snackbar']
+        });
+      }
     });
   }
 
@@ -166,16 +163,5 @@ export class ProductlistComponent implements OnInit {
 
   addToShopping(product: Product): void {
     this.productClicked.emit(product);
-  }
-
-  handleFileInputChange(l: FileList): void {
-    this.file_store = l;
-    if (l.length) {
-      const f = l[0];
-      const count = l.length > 1 ? `(+${l.length - 1} files)` : '';
-      this.display.patchValue(`${f.name}${count}`);
-    } else {
-      this.display.patchValue('');
-    }
   }
 }
